@java8 stream practice

+++ Java 8 Practice/src/practice/java8/StreamLambda.java
@@ -50,7 +50,7 @@ public class StreamLambda {
 		list.add(new Product(4,"One Plus8",32000f));
 		
 		
-		double totalPrice = list.stream().mapToDouble(p->p.price).sum();
+		float totalPrice = (float)list.stream().mapToDouble(p->p.price).sum();
 		System.out.println(totalPrice);
 	}
